LayoutTests/media/media-source/media-source-real-play-after-eos.html

<!DOCTYPE html>
<html>
    <head>
        <title>media-source-real-play-after-eos (inspired from media-source-real-seek-and-play)</title>
        <script src="mp4-generator.js"></script>
        <script src="../video-test.js"></script>
        <script>
         // Tests sequential append of contiguous segments with real MP4 data.

         // Verifies that playback of a second segment works after playback of the first segment
         // ended with an EOS event at the platform level.

         var source;
         var videoSourceBuffer;

         async function runTest() {
             findMediaElement();

             if (!MediaSource.isTypeSupported(MP4.VIDEO_TYPE)) {
                 failTest(`${MP4.VIDEO_TYPE} is not supported`);
                 return;
             }

             source = new MediaSource();
             video.src = URL.createObjectURL(source);
             await waitFor(source, 'sourceopen');

             videoSourceBuffer = source.addSourceBuffer(MP4.VIDEO_TYPE);

             var videoInitData = MP4.initSegment({
                 timescale: 1000,
                 tracks: [{ id: 1, type: 'video', timescale: 1000 }]
             });

             // First segment: 4 seconds
             var videoSegment1 = MP4.mediaSegment({
                 sequenceNumber: 1,
                 tracks: [{ id: 1, type: 'video', baseDecodeTime: 0, samples: [
                     { duration: 1000, compositionTimeOffset: 0, isSync: true },
                     { duration: 1000, compositionTimeOffset: 0, isSync: false },
                     { duration: 1000, compositionTimeOffset: 0, isSync: false },
                     { duration: 1000, compositionTimeOffset: 0, isSync: false },
                 ]}]
             });

             // Second segment: 4 more seconds (contiguous)
             var videoSegment2 = MP4.mediaSegment({
                 sequenceNumber: 2,
                 tracks: [{ id: 1, type: 'video', baseDecodeTime: 4000, samples: [
                     { duration: 1000, compositionTimeOffset: 0, isSync: true },
                     { duration: 1000, compositionTimeOffset: 0, isSync: false },
                     { duration: 1000, compositionTimeOffset: 0, isSync: false },
                     { duration: 1000, compositionTimeOffset: 0, isSync: false },
                 ]}]
             });

             videoSourceBuffer.appendBuffer(videoInitData);
             await waitFor(videoSourceBuffer, 'updateend');

             videoSourceBuffer.appendBuffer(videoSegment1);
             await waitFor(videoSourceBuffer, 'updateend');

             run('source.endOfStream()');
             testExpected('source.readyState', 'ended');

             // Play the first segment until EOS.
             video.play();
             video.currentTime = 3.5;
             await waitFor(video, 'seeked');
             await waitFor(video, 'ended');

             videoSourceBuffer.appendBuffer(videoSegment2);
             await waitFor(videoSourceBuffer, 'updateend');

             run('source.endOfStream()');
             testExpected('source.readyState', 'ended');
             testExpectedEqualWithTolerance('source.duration', 8, 0.01);

             // Play the second segment, the seek towards its end should work as well.
             video.play();
             video.currentTime = 7.5;
             await waitFor(video, 'seeked');
             await waitFor(video, 'ended');

             endTest();
         }
        </script>
    </head>
    <body onload="runTest()">
        <video></video>
    </body>
</html>
