LayoutTests/media/media-source/media-source-real-play-after-eos-expected.txt


EVENT(sourceopen)
EVENT(updateend)
EVENT(updateend)
RUN(source.endOfStream())
EXPECTED (source.readyState == 'ended') OK
EVENT(seeked)
EVENT(ended)
EVENT(updateend)
RUN(source.endOfStream())
EXPECTED (source.readyState == 'ended') OK
EXPECTED (source.duration == '8') OK
EVENT(seeked)
EVENT(ended)
END OF TEST
